+++ extensions/linear/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Linear Changelog
 
+## [Add attachments to Create Issue command] - 2023-01-24
+
+- Add a file picker on the `Create Issue` command to add a single attachment on a newly created Linear issue
+
 ## [Add due date action] - 2023-01-19
 
 - Add a new action to set due dates on issues